EDA221_Assignment1: Set proper type for window's parameters
Fixes: 42ee685a63e03bc086cbeceadf7c68ae471e3e27 ("Remove templated
variables as VS in the labs does not support them")

+++ src/EDA221_Assignment1/SolarSystem.cpp
@@ -34,9 +34,9 @@
 
 namespace config
 {
-	constexpr float msaa_rate    =    1u;
-	constexpr float resolution_x = 1600u;
-	constexpr float resolution_y =  900u;
+	constexpr unsigned int msaa_rate    =    1u;
+	constexpr unsigned int resolution_x = 1600u;
+	constexpr unsigned int resolution_y =  900u;
 }
 
 enum class shader_bindings {
